Flask/requirements.txt
Flask
mysql-connector-python
waitress
pyyaml==5.3.1
docker-compose
